+++ README.md
@@ -1,5 +1,18 @@
 # DICOM parser in Go [![GoDoc](https://godoc.org/github.com/gillesdemey/go-dicom?status.svg)](https://godoc.org/github.com/gillesdemey/go-dicom) [![wercker status](https://app.wercker.com/status/c250d72bc82a5d8f267c7ee0b9e839bc/s/master "wercker status")](https://app.wercker.com/project/bykey/c250d72bc82a5d8f267c7ee0b9e839bc)
 
+This is a fork of github.com/gillesdemey/go-dicom. Changes are:
+
+- Many bug fixes, especially around handling of sequences.
+- Simplify the API. All the functions are synchronous.
+- Better library supports around tags & uids.
+- Rudimentary support for writing DICOM files. This part is not complete yet.
+- Adds fuzz tests and tests that ensure compatibility with pydicom.
+
+TODO:
+- Non-utf8 coding system support.
+- A multi-image file. Functionality is almost there, but I haven't had time to complete it.
+- Native pixeldata format. It'll be parsed as just []byte.
+
 ## Usage
 ```Go
 package main
