+++ dicom.go
@@ -24,7 +24,6 @@ package dicom
 import (
 	"bytes"
 	"encoding/binary"
-	"errors"
 	"fmt"
 	"io"
 )
@@ -92,6 +91,8 @@ func Parse(in io.Reader, bytes int64) (*DicomFile, error) {
 		return nil, buffer.Error()
 	}
 	file := &DicomFile{Elements: metaElems}
+
+	// Change the transfer syntax for the rest of the file.
 	elem, err := file.LookupElement("TransferSyntaxUID")
 	if err != nil {
 		return nil, err
@@ -100,16 +101,15 @@ func Parse(in io.Reader, bytes int64) (*DicomFile, error) {
 	if err != nil {
 		return nil, err
 	}
-	// read endianness and explicit VR
-	endianess, implicit, err := ParseTransferSyntaxUID(transferSyntaxUID)
+	endian, implicit, err := ParseTransferSyntaxUID(transferSyntaxUID)
 	if err != nil {
 		return nil, err
 	}
-	// modify buffer according to new TransferSyntaxUID
-	buffer.bo = endianess
-	buffer.implicit = implicit
+	buffer.PushTransferSyntax(endian, implicit)
+	defer buffer.PopTransferSyntax()
 
-	for buffer.Len() != 0 && buffer.Error() == nil {
+	// Now read the list of elements.
+	for buffer.Len() != 0 {
 		elem := ReadDataElement(buffer)
 		if buffer.Error() != nil {
 			break
@@ -117,89 +117,6 @@ func Parse(in io.Reader, bytes int64) (*DicomFile, error) {
 		file.Elements = append(file.Elements, *elem)
 	}
 	return file, buffer.Finish()
-}
-
-// Consume the DICOM magic header and metadata elements from a Dicom
-// file. Errors are reported through d.Error().
-func ParseFileHeader(d *Decoder) []DicomElement {
-	d.PushTransferSyntax(binary.LittleEndian, ExplicitVR)
-	defer d.PopTransferSyntax()
-	d.Skip(128) // skip preamble
-
-	// check for magic word
-	if s := d.DecodeString(4); s != "DICM" {
-		d.SetError(errors.New("Keyword 'DICM' not found in the header"))
-		return nil
-	}
-
-	// (0002,0000) MetaElementGroupLength
-	metaElem := ReadDataElement(d)
-	if d.Error() != nil {
-		return nil
-	}
-	if metaElem.Tag != TagMetaElementGroupLength {
-		d.SetError(fmt.Errorf("MetaElementGroupLength not found; insteadfound %s", metaElem.Tag.String()))
-	}
-	metaLength, err := GetUInt32(*metaElem)
-	if err != nil {
-		d.SetError(fmt.Errorf("Failed to read uint32 in MetaElementGroupLength: %v", err))
-		return nil
-	}
-	if d.Len() <= 0 {
-		d.SetError(fmt.Errorf("No data element found"))
-		return nil
-	}
-	metaElems := []DicomElement{*metaElem}
-
-	// Read meta tags
-	d.PushLimit(int64(metaLength))
-	defer d.PopLimit()
-	for d.Len() > 0 {
-		elem := ReadDataElement(d)
-		if d.Error() != nil {
-			break
-		}
-		metaElems = append(metaElems, *elem)
-	}
-	return metaElems
-}
-
-// Inverse of ParseFileHeader. Errors are reported via e.Error()
-func WriteFileHeader(e *Encoder,
-	transferSyntaxUID string,
-	sopClassUID string,
-	sopInstanceUID string) {
-	e.PushTransferSyntax(binary.LittleEndian, ExplicitVR)
-	defer e.PopTransferSyntax()
-	encodeSingleValue := func(encoder *Encoder, tag Tag, v interface{}) {
-		elem := DicomElement{
-			Tag:   tag,
-			Vr:    "", // autodetect
-			Vl:    1,
-			Value: []interface{}{v},
-		}
-		EncodeDataElement(encoder, &elem)
-	}
-
-	// Encode the meta info first.
-	subEncoder := NewEncoder(binary.LittleEndian, ExplicitVR)
-	encodeSingleValue(subEncoder, TagFileMetaInformationVersion, []byte("0 1"))
-	encodeSingleValue(subEncoder, TagTransferSyntaxUID, transferSyntaxUID)
-	encodeSingleValue(subEncoder, TagMediaStorageSOPClassUID, sopClassUID)
-	encodeSingleValue(subEncoder, TagMediaStorageSOPInstanceUID, sopInstanceUID)
-	encodeSingleValue(subEncoder, TagImplementationClassUID, DefaultImplementationClassUID)
-	encodeSingleValue(subEncoder, TagImplementationVersionName, DefaultImplementationVersionName)
-	// TODO(saito) add more
-	metaBytes, err := subEncoder.Finish()
-	if err != nil {
-		e.SetError(err)
-		return
-	}
-
-	e.EncodeZeros(128)
-	e.EncodeString("DICM")
-	encodeSingleValue(e, TagMetaElementGroupLength, uint32(len(metaBytes)))
-	e.EncodeBytes(metaBytes)
 }
 
 func doassert(x bool) {
